rename {SpectralProcessor => spectral_processor}.hpp
