+++ include/KernelSmoother.hpp
@@ -11,7 +11,7 @@
 
 #include "Allocator.hpp"
 #include "SIMDSupport.hpp"
-#include "SpectralProcessor.hpp"
+#include "spectral_processor.hpp"
 #include "table_reader.hpp"
 
 template <typename T, typename Allocator = aligned_allocator, bool auto_resize_fft = false>
