+++ include/spectral_processor.hpp
